scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
openai==1.3.0
python-dotenv==1.0.0 